Moved ProjectFileInfo out in a separate file
